fix int min

+++ lib/printk/itoa_stack_max.c
@@ -50,7 +50,7 @@ int itoa_stack_max(char *nbr_buff, int is_unsigned, long int nb, const char *cha
 
 		const size_t new_nb = (-nb / len_charset);
 		if (new_nb > 0)
-			itoa_stack_max(nbr_buff + 1, 1, (-nb / len_charset), char_set);
+			itoa_stack_max(nbr_buff + 1, 1, -(nb / (int)len_charset), char_set);
 
 		last_write = itoa_stack_max_len(nbr_buff);
 		nbr_buff[last_write++] = char_set[(-(nb % (int)len_charset) + len_charset) % (int)len_charset];
